app/RecognitionService/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
pillow
numpy
pandas
python-multipart
easyocr

prometheus-client
prometheus-fastapi-instrumentator